+++ CHANGELOG.md
@@ -9,6 +9,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 - MultipartRestClient in test-utils: client for multipart file uploads
 - LiquibaseDatabaseBootstrapper in test-utils: bootstraps a database using your liquibase scripts 
 
+### Fixed
+- Generate random ZonedDateTime in different time zones and provide option to generate in UTC specific timezone
+
 ## [1.4.1] - 2017-02-13
 
 ### Fixed